+++ changelog.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Boosts may now completely be left out
+- Console commands now use the configured connection
 
 ### Changed
 - BoolQueries can now also be nested